requirements.txt
sqlalchemy
Levenshtein~=0.27.1
psycopg2-binary~=2.9.10
dotenv~=0.9.9
python-dotenv~=1.1.0
python-telegram-bot~=21.10
Telethon~=1.38.1
requests~=2.32.3